Update instructions
Updated instructions per referee request to refer to one stock.

Note: the instructions are only appropriate for Session III of the game.
They need to be further customized for each Session. For Session I and II, one needs to cut the gamification elements we don't use. For Session IV, we additionally need to change the transition probabilities.